feat(battery)!: Switch to fuel gauge API
ZMK now uses the fuel gauge API instead of the sensor API for reading
battery state of charge. The zmk,battery-nrf-vddh and zmk,maxim-max17048
drivers have been rewritten to use the fuel gauge API.

There is also a new zmk,fuel-gauge-sensor driver which can be used to
wrap any Zephyr driver that uses the old sensor API, as well as a new
zmk,fuel-gauge-voltage driver which estimates state of charge by reading
from a battery voltage sensor.

This is a breaking change if your board uses "voltage-divider" or
"zmk,battery-voltage-divider" as the zmk,battery chosen node. To
integrate this change:

 1. Replace 'compatible = "zmk,battery-voltage-divider"' with
    'compatible = "voltage-divider"'.

 2. Add a new "zmk,fuel-gauge-voltage" node with the "voltage-divider"
    node as its "sensor" property. If the voltage divider node is
    labeled "vbatt", then it would look like this:

    vbatt_gauge: vbatt_gauge {
        compatible = "zmk,fuel-gauge-voltage";
        sensor = <&vbatt>;
    };

 3. Change the "zmk,battery" property in the "chosen" node to use the
    new "zmk,fuel-gauge-voltage" node.

    chosen {
        zmk,battery = &vbatt_gauge;
    };

This is also a breaking change if your board uses a "maxim,max17055",
"maxim,max17262", or "ti,bq274xx" fuel gauge or a custom driver that
uses the sensor API. To integrate this change, you can wrap it with a
"zmk,fuel-gauge-sensor" node:

    chosen {
        zmk,battery = &fuel_gauge;
    };

    fuel_gauge: fuel_gauge {
        compatible = "zmk,fuel-gauge-sensor";
        sensor = <&fuel_gauge_sensor>;
    };

    fuel_gauge_sensor: bq274xx@55 {
        compatible = "ti,bq274xx";
        ...
    };